feat: `browser` no longer `enables` by default
__*BREAKING CHANGE*__

In the past we enbaled anyting in the `DEBUG` localStorage key,
this lead to some undesired behaviours when working with sandboxed
iframes.

Now we will let the you control this completely, as from what I can see
the majority of folk I see, use it anyway.

Still makes sense to run this for Node applications, which is left un
touched.

If you want the existing behaviour, just attached this before you start
logging.

```ts
import { enable } from 'diary';
enable(localStorage.DEBUG || 'a^');
```

fixes: #15